handling response and deleting customers

+++ src/main/java/org/example/service/DefaultCustomerService.java
@@ -3,7 +3,10 @@ package org.example.service;
 import org.example.model.Customer;
 import org.example.repository.CustomerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
@@ -26,13 +29,21 @@ public class DefaultCustomerService implements CustomerService {
   }
 
   @Override
-  public Customer getCustomer(Long id) {
+  public ResponseEntity<Customer> getCustomer(Long id) {
     Optional<Customer> optionalCustomer = repository.findById(id);
-    return optionalCustomer.orElseGet(Customer::new);
+    if(optionalCustomer.isPresent()) {
+      return new ResponseEntity<>(optionalCustomer.get(), HttpStatus.OK);
+    }
+    throw new ResponseStatusException(HttpStatus.NOT_FOUND, "customer not found");
   }
 
   @Override
   public void deleteCustomer(Long id) {
+    // before deleting, verify that the Customer exists in the DB
+    ResponseEntity<Customer> customer = getCustomer(id);
+    if(customer.getStatusCode().is4xxClientError()) {
+      throw new ResponseStatusException(HttpStatus.NOT_FOUND, "customer not found");
+    }
     repository.deleteById(id);
   }
